# backend/docs/swagger.yaml
openapi: 3.0.0
info:
  title: E-commerce Clothes API
  version: 1.0.0
  description: API documentation for E-commerce Clothes application
  contact:
    name: API Support
    email: support@example.com
servers:
  - url: http://localhost:3000
    description: Development server
  - url: https://your-production-url.com
    description: Production server
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    Product:
      type: object
      properties:
        _id:
          type: string
        name:
          type: string
        price:
          type: number
        discountPrice:
          type: number
        description:
          type: string
        countInStock:
          type: integer
        category:
          type: string
        brand:
          type: string
        sizes:
          type: array
          items:
            type: string
        colors:
          type: array
          items:
            type: string
        collections:
          type: string
        material:
          type: string
        gender:
          type: string
          enum: [men, women, unisex]
        images:
          type: array
          items:
            type: string
        isFeatured:
          type: boolean
        isPublished:
          type: boolean
        tags:
          type: array
          items:
            type: string
        dimensions:
          type: object
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
...
    Order:
      type: object
      properties:
        _id:
          type: string
        user:
          type: string
        items:
          type: array
          items:
            type: object
        shippingAddress:
          type: object
          properties:
            fullName:
              type: string
            address:
              type: string
            city:
              type: string
            postalCode:
              type: string
            country:
              type: string
        paymentMethod:
          type: string
        paymentResult:
          type: object
        totalPrice:
          type: number
        isPaid:
          type: boolean
        paidAt:
          type: string
          format: date-time
        isDelivered:
          type: boolean
        deliveredAt:
          type: string
          format: date-time
        status:
          type: string
          enum: [pending, processing, shipped, delivered, cancelled]
    Error:
      type: object
      properties:
        message:
          type: string
        error:
          type: string
...
paths:
...
  /api/products:
    get:
      tags: [Products]
      summary: Get all products with optional filters
      parameters:
        - in: query
          name: collection
          schema:
            type: string
          description: Filter by collection
        - in: query
          name: category
          schema:
            type: string
          description: Filter by category
        - in: query
          name: brand
          schema:
            type: string
          description: Filter by brand (comma-separated)
        - in: query
          name: size
          schema:
            type: string
          description: Filter by size (comma-separated)
        - in: query
          name: color
          schema:
            type: string
          description: Filter by color
        - in: query
          name: gender
          schema:
            type: string
          description: Filter by gender (men/women)
        - in: query
          name: minPrice
          schema:
            type: number
          description: Minimum price filter
        - in: query
          name: maxPrice
          schema:
            type: number
          description: Maximum price filter
        - in: query
          name: page
          schema:
            type: integer
            default: 1
          description: Page number
        - in: query
          name: limit
          schema:
            type: integer
            default: 12
          description: Number of products per page
        - in: query
          name: sort
          schema:
            type: string
            enum: [price_asc, price_desc, name_asc, name_desc, newest, oldest]
          description: Sort products
      responses:
        '200':
          description: List of products
          content:
            application/json:
              schema:
                type: object
                properties:
                  products:
                    type: array
                    items:
                      $ref: '#/components/schemas/Product'
                  totalProducts:
                    type: integer
                  totalPages:
                    type: integer
                  currentPage:
                    type: integer
        '500':
          description: Server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

    post:
      tags: [Products]
      summary: Create a new product (Admin only)
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - name
                - price
                - description
                - countInStock
                - category
                - brand
              properties:
                name:
                  type: string
                price:
                  type: number
                discountPrice:
                  type: number
                description:
                  type: string
                countInStock:
                  type: integer
                category:
                  type: string
                brand:
                  type: string
                sizes:
                  type: array
                  items:
                    type: string
                colors:
                  type: array
                  items:
                    type: string
                collections:
                  type: string
                material:
                  type: string
                gender:
                  type: string
                images:
                  type: array
                  items:
                    type: string
                isFeatured:
                  type: boolean
                isPublished:
                  type: boolean
                tags:
                  type: array
                  items:
                    type: string
      responses:
        '201':
          description: Product created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Product'
        '401':
          description: Unauthorized
        '403':
          description: Admin access required
...
  /api/products/{id}:
    get:
      tags: [Products]
      summary: Get product by ID
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
          description: Product ID
      responses:
        '200':
          description: Product details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Product'
        '404':
          description: Product not found

    put:
      tags: [Products]
      summary: Update product (Admin only)
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
          description: Product ID
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Product'
      responses:
        '200':
          description: Product updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Product'
        '404':
          description: Product not found
        '401':
          description: Unauthorized
        '403':
          description: Admin access required

    delete:
      tags: [Products]
      summary: Delete product (Admin only)
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
          description: Product ID
      responses:
        '200':
          description: Product deleted successfully
        '404':
          description: Product not found
        '401':
          description: Unauthorized
        '403':
          description: Admin access required
  /api/products/similar/{id}:
    get:
      tags: [Products]
      summary: Get similar products
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
          description: Product ID
      responses:
        '200':
          description: List of similar products
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Product'

  # Cart Routes
...
  /api/orders/my-orders:
    get:
      tags: [Orders]
      summary: Get user's orders
      security:
        - bearerAuth: []
      responses:
        '200':
          description: List of user's orders
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Order'
  /api/orders/{orderId}:
    get:
      tags: [Orders]
      summary: Get order by ID
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: orderId
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Order details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Order'
        '404':
          description: Order not found

  # Upload Routes
...
  /api/admin/products:
    get:
      tags: [Admin - Products]
      summary: Get all products for admin (Admin only)
      security:
        - bearerAuth: []
      responses:
        '200':
          description: List of all products
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Product'
  /api/admin/products/{id}:
    put:
      tags: [Admin - Products]
      summary: Update product (Admin only)
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Product'
      responses:
        '200':
          description: Product updated successfully

    delete:
      tags: [Admin - Products]
      summary: Delete product (Admin only)
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Product deleted successfully

  # Admin Routes - Orders
...
  /api/admin/orders/{id}:
    get:
      tags: [Admin - Orders]
      summary: Get order by ID (Admin only)
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Order details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Order'

    put:
      tags: [Admin - Orders]
      summary: Update order status (Admin only)
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                status:
                  type: string
                  enum: [pending, processing, shipped, delivered, cancelled]
                isDelivered:
                  type: boolean
                deliveredAt:
                  type: string
                  format: date-time
      responses:
        '200':
          description: Order updated successfully